tests/wpt/tests/css/css-scroll-anchoring/abspos-contributes-to-static-parent-bounds.html

<!DOCTYPE html>
<meta name="viewport" content="width=device-width,initial-scale=1">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>

body, html, #static { height: 0; }
#abs {
  position: absolute;
  left: 50px;
  top: 50px;
  height: 1200px;
  padding: 50px;
  border: 5px solid gray;
}
#anchor {
  background-color: #afa;
  width: 100px;
  height: 100px;
}

</style>
<div id="static">
  <div id="abs">
    <div id="changer"></div>
    <div id="anchor"></div>
  </div>
</div>
<script>

// Tests that the "bounds" of an element, for the purpose of visibility in the
// anchor node selection algorithm, include any space occupied by the element's
// positioned descendants.

test(() => {
  document.scrollingElement.scrollTop = 120;
  document.querySelector("#changer").style.height = "100px";
  assert_equals(document.scrollingElement.scrollTop, 220);
}, "Abs-pos with zero-height static parent.");

</script>
